HYDRA-600 : Add filtering scene index interface to flow viewport tool… (#21)
* HYDRA-600 : Add filtering scene index interface to flow viewport toolkit.

* HYDRA-600 : Fix OSX compilation.

* HYDRA-600 : Fixes from the code review.

* HYDRA-600 : more modifications from code review.

* HYDRA-600 :; found a potential issue with declaring ref ptr

* HYDRA-600 : switch to using vector instead set for ViewportInformationAndSceneIndicesPerViewportData and add const where it was possible.

+++ test/lib/mayaUsd/render/mayaToHydra/cpp/testMayaSceneFlattening.cpp
@@ -32,7 +32,7 @@ TEST(MayaSceneFlattening, childHasFlattenedTransform)
 {
     // Setup inspector for the first scene index
     const SceneIndicesVector& sceneIndices = GetTerminalSceneIndices();
-    ASSERT_GT(sceneIndices.size(), static_cast<size_t>(0));
+    ASSERT_GT(sceneIndices.size(), 0u);
     SceneIndexInspector inspector(sceneIndices.front());
 
     // Retrieve the child cube prim
@@ -47,7 +47,7 @@ TEST(MayaSceneFlattening, childHasFlattenedTransform)
         return parentIsCube;
     };
     PrimEntriesVector foundPrims = inspector.FindPrims(findCubePrimPredicate, 1);
-    ASSERT_EQ(foundPrims.size(), static_cast<size_t>(1));
+    ASSERT_EQ(foundPrims.size(), 1u);
     HdSceneIndexPrim cubePrim = foundPrims.front().prim;
 
     // Extract the Hydra xform matrix from the cube prim
